refactor(ui): Change directory name of bulk conclusion table
