+++ apps/clearance_ui/src/components/bulk_conclusion_table/columns.tsx
